chore(runtime): move auth to a subfolder
