+++ packages/@aws-cdk/aws-bedrock-agentcore-alpha/lib/index.ts
@@ -14,7 +14,7 @@ export * from './runtime/perms';
 export * from './runtime/types';
 export * from './runtime/runtime-base';
 export * from './runtime/runtime-artifact';
-export * from './runtime/runtime-authorizer-configuration';
+export * from './runtime/inbound-auth/runtime-authorizer-configuration';
 export * from './runtime/runtime-endpoint-base';
 export * from './runtime/runtime-endpoint';
 export * from './runtime/runtime';
